StaticValidator check for zero-arity rules and sentences.
Also found in the process that the GdlCleaner sometimes needs to run
multiple times, so fixing that. The GdlCleaner sometimes results in
GdlPropositions scattered around, which is technically wrong but not a
real problem, so making that a "warning" in the validator for now.
Long-term, the GdlCleaner should fix that, too.

Fixing related tests and test classes.