Improve validator warning handling; add new warning to StaticValidator.
Warnings can now be accumulated and recorded, then shown to the user,
instead of being println'd immediately.

The new StaticValidator warning checks for the use of a constant as both
a sentence and function in a game description. This isn't clearly
prohibited by the GDL specification, but is generally a bad idea.