Fix arbitrary recursion handling in the ProverStateMachine.
The old approach was a hack that worked for one particular game I was
writing, but did not give correct results in other cases. (The approach
before that resulted in StackOverflowErrors.) The new approach should
give correct results in all cases.

It does not appear there are any performance regressions for existing
games.