Don't rotate the player assigned to the role 'random' when generating multiple matches
The current code rotates all players when generating multiple matches
for the same game. This change leaves the player associated with the
'random' role at the same index in the player list and rotates all the
other players. If there is no player associated with the 'random' role,
the code works as before.