Partially split out state-machines project.
This is meant to include the Prover and ProverStateMachine as well
as the StateMachine and related interfaces.

Gdl utils should end up in an upstream project.

Some errors remain around the GamerLogger, which requires the
Match class. There must be some better solution there, maybe
something involving passing in some log interface.

Propnet stuff will probably go in a separate project downstream,
so as to not require it for someone who just wants the
ProverStateMachine.