Make the pre-scheduling player status check case insensitive
When the game server displays the status color bubble thing in the UI it uses
`.toLowerCase()` to allow for uppercase statuses.  But when it checks whether it
can actually schedule a requested match, it compares the raw status string
against a lowercase version, and so it'll just hang forever on "pending" and
never schedule the match with players that report their status in uppercase.